Confidence Intervals for Generalized Kumaraswamy Regression Parameters

Computes confidence intervals for model parameters in fitted gkwreg objects using Wald (normal approximation) method based on asymptotic theory.

Usage

# S3 method for class 'gkwreg'
confint(object, parm, level = 0.95, ...)

Arguments

object

An object of class "gkwreg" from gkwreg.

parm

A specification of which parameters are to be given confidence intervals, either a vector of numbers or a vector of names. If missing, all parameters are considered.

level

The confidence level required. Default is 0.95.

...

Additional arguments (currently unused).

Value

A matrix (or vector) with columns giving lower and upper confidence limits for each parameter. These will be labeled as (1-level)/2 and 1 - (1-level)/2 in percent (by default 2.5 percent and 97.5 percent).

Details

The confidence intervals are computed using the Wald method based on asymptotic normality of maximum likelihood estimators: $$CI = \hat{\theta} \pm z_{\alpha/2} \times SE(\hat{\theta})$$ where \(z_{\alpha/2}\) is the appropriate normal quantile and \(SE(\hat{\theta})\) is the standard error from the Hessian matrix.

The model must have been fitted with hessian = TRUE (the default) in gkw_control. If standard errors are not available, an error is raised.

See also

Author

Lopes, J. E.

Examples

# \donttest{
data(GasolineYield)
fit <- gkwreg(yield ~ batch + temp, data = GasolineYield, family = "kw")
#> Warning: NaNs produced

# 95 percent confidence intervals
confint(fit)
#> Warning: some standard errors are NA or infinite; intervals may be unreliable
#>                         2.5 %     97.5 %
#> alpha:(Intercept)  0.46546878  0.5566185
#> alpha:batch1       0.79054008  0.8996843
#> alpha:batch2       0.53953733  0.6574229
#> alpha:batch3       0.65732574  0.7753480
#> alpha:batch4       0.41885996  0.5292772
#> alpha:batch5       0.46088005  0.5826205
#> alpha:batch6       0.39671081  0.5169558
#> alpha:batch7       0.17469502  0.2864693
#> alpha:batch8       0.14465787  0.2682429
#> alpha:batch9       0.07755916  0.2148541
#> alpha:temp                NaN        NaN
#> beta:(Intercept)  24.27478345 25.6466010

# 90 percent confidence intervals
confint(fit, level = 0.90)
#> Warning: some standard errors are NA or infinite; intervals may be unreliable
#>                           5 %       95 %
#> alpha:(Intercept)  0.47279602  0.5492913
#> alpha:batch1       0.79931383  0.8909105
#> alpha:batch2       0.54901378  0.6479465
#> alpha:batch3       0.66681317  0.7658606
#> alpha:batch4       0.42773605  0.5204011
#> alpha:batch5       0.47066637  0.5728342
#> alpha:batch6       0.40637691  0.5072897
#> alpha:batch7       0.18368019  0.2774841
#> alpha:batch8       0.15459247  0.2583083
#> alpha:batch9       0.08859585  0.2038174
#> alpha:temp                NaN        NaN
#> beta:(Intercept)  24.38505944 25.5363251

# Specific parameters
confint(fit, parm = "alpha:(Intercept)")
#> Warning: some standard errors are NA or infinite; intervals may be unreliable
#>                       2.5 %    97.5 %
#> alpha:(Intercept) 0.4654688 0.5566185
confint(fit, parm = 1:3)
#> Warning: some standard errors are NA or infinite; intervals may be unreliable
#>                       2.5 %    97.5 %
#> alpha:(Intercept) 0.4654688 0.5566185
#> alpha:batch1      0.7905401 0.8996843
#> alpha:batch2      0.5395373 0.6574229
# }